requirements.txt
RPi.GPIO==0.7.1
paho-mqtt==2.0.0
adafruit-circuitpython-dht==3.7.9
python-dotenv==1.0.1
